# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "description": "API responsável por gerenciar os filmes do MovieFlix.",
        "version": "1.0.0"
    },

    "paths": {
        "/movies": {
            "get": {
                "tags": ["Filmes"],
                "summary": "Listar todos os filmes",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Filmes"],
                "summary": "Cadastrar um filme",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": ["Filmes"],
                "summary": "Atualizar um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser atualizado.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PutError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Filmes"],
                "summary": "Remover um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser removido.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme removido com sucesso."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteError"
                                }
                            }
                        }
                    }
                }
            }
        },

        "/movies/{genreName}": {
            "get": {
                "tags": ["Filmes"],
                "summary": "Listar filmes de um determinado gênero",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Nome do gênero dos filmes.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilterError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },

    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme."
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme."
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Genero do filme."
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero."
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero."
                            }
                        }  
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme."
                    },
                    "language": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do idioma."
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do idioma."
                            }
                        }  
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de oscar ganhos pelo filme."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lancamento do filme."
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme."
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme."
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme."
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de oscar ganhos pelo filme."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lancamento do filme."
                    }
                }
            },
            "NotFound": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Filme não encontrado."
                    }
                }
            },
            "PostError": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Falha ao cadastrar filme."
                    }
                }
            },
            "PutError": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Falha ao atualizar o registro do filme."
                    }
                }
            },
            "DeleteError": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Falha ao deletar filme."
                    }
                }
            },
            "FilterError": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Falha ao filtrar filmes por gênero."
                    }
                }
            }
        }
    }
}